requirements.txt
asgiref==3.6.0
Django==4.1.4
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2022.7
sqlparse==0.4.3
django-modeltranslation~=0.18.7
